Refactor and improve definition list
The parser of the definition list was quite limited, it didn't support
all feature that a definition list should support, and didn't fit
in the new node structure.

By changing the implementation, upgrading the test result to cover
the sphinx output this new rule is much closer to what we expect
according to the doc-utils spec and sphinx output.