Fixes ref role links
The text role `ref` is now supported as it should.
Simplyfied the way links are rendered without markup in the titles.
This allows us again to create links using just text.

There is a major issue with using the subparsers in this change
the sub documents are isolated from the main document. This causes
issues handling includes. We should refactor this in another effort.