[TASK] Make configuration less complex
By allowing the inventories at the root of our configuration we
are more following the standard way of configuring an application
in symfony. This will help us to support other configuration formats
if we want to.

By moving the initialization of the inventory repository to a parameter
extensions can not provide inventories. This could be usefull for non
sphinx documentation sets, but still using the interlink functionality.
For example when we create a extension that links to the php documentation.