[FIX] support single document rendering
To make it possible to extend the html rendering with a single document
we should allow customization of the html rending stack.
The easiest way to do this this to add a new renderer that is
reusing the delegating renderer from the html part. By allowing
extensions to define their own delegating renderer with a factory
set, we can switch towards another stack.

This is a rather complex use case, which we do not support out of the
box but needed for typo3.