[BUGFIX] Use output file path for relative url calculations
In standard renderings output and input filenames with their paths are the same except for the file ending. However, in project based rendering nodes the currentFileName in the rendering context is null, leading to an error.

In these cases we add the outputfile path manually to the render context.

The relative urls must now be calculated relative to the output path, not the input path.