[FEATURE] Handle Global Menu depth independently from toc
If the toctree on the root page is limited to a certain level count, like `:maxdepth: 1` still the complete page tree is displayed in Sphinx.

However as not ever project needs a full document tree and the creation of the document tree is time-consuming, an option for the guides.xml, max-menu-depth, was introduced. `max-menu-depth = 0` disables the global menu all together.