Introduce CompilerContextInterface
The Compiler context is a very important class in our compiler, however
extending projects may want to extend the information caried around with the context
to add extra information for specific use cases in custom compiler passes.

To make this possible I introduced a new interface to be able to wrap the compilercontext
and extend the functionality. Changing the interfaces would be a backward compatibility break
so we need to introduce this step by step.