[FEATURE] Make index file configurable
The index file will also be used in automatic menu
building.

As oposed to setting "input-file" not only the one file mentioned will be rendererd but all files from the directory. However the name of the input file can be changed.

Some markdown projects like to change the name of the index file. For example powermail is using
Readme.md as index file in every directory:

https://github.com/in2code-de/powermail/tree/master/Documentation

References #1108